refactor: move tracer.js to lib/tracer.js
